refactor: rename `MessageListView` -> `ChatView`
We want to unite it with the chat header.
And move into a dir so that we can add `styles.module.scss` for it.
