+++ packages/frontend/src/components/screens/MainScreen/MainScreen.tsx
@@ -21,7 +21,7 @@ import { BackendRemote, onDCEvent, Type } from '../../../backend-com'
 import Button from '../../Button'
 import Icon from '../../Icon'
 import SearchInput from '../../SearchInput'
-import MessageListView from '../../MessageListView'
+import MessageListView from '../../ChatView/ChatView'
 import useChat from '../../../hooks/chat/useChat'
 import useDialog from '../../../hooks/dialog/useDialog'
 import useKeyBindingAction from '../../../hooks/useKeyBindingAction'
